256*256 crops produced successfully
The corruptions caused by the previous commit were from slicing the np
array out of bounds.  The array is now padded before clipping to avoid
this and has the effect of wrap-around if the ROI is too close to the
edge of the image.  Code that calculates average ROI size has been
added. ImageViewer works well enough.